Fix Crash
Fixed the crashing of the channel. It was a combination of two things. In the `GetStreamURL` function there was unexpected whitespace in front of `best_bandwidth` and `best_url` lines.

It looks like 9 News KUSA has moved their stream because the channel is no longer able to find it, so I commented the info for Channel 9 out of the app. It now works well with these three channels.

If anybody knows how to get the m3u8 url for the channel 9 stream please let me know!